Change annoy-a-tron to be a runner
Instead of having each script implement its own randomness, change the
`annoy-a-tron` script to be a runner script that implements its the
randomness, and then shells out to whatever script is currently
'active'. This should simplify the script writing and testing for
contributors.

Additionally, I reduced the chance of the annoying a little, and udpated
`random-time.py` to include "AM" or "PM" in the announcement.