Run a random script every time
Instead of running a pre-determined script each time, start running a
randomly chosen one from the list of available scripts. This should make
the annoying stuff a little less obvious.

I also removed the `no-op.sh` (since it doesn't make sense with this new
pick-a-random-one idea), and removed `random-pbcopy.sh` since it has a
habit of failing randomly and in weird ways.